Added guess functionality (see details)
When you click guess, popup asking what the picture is. If you click cancel or enter an empty string, nothing happens. If you enter a guess, one of three things will happen: either 1) the guess was correct and a popup confirms that, 2) the guess was incorrect and a popup confirms that, or 3) the guess was close enough to being correct to trigger another dialog box asking 'did you mean <answer>?'. If the user says 'yes', the correct guess popup comes up.