Category info and parser added
The CategoryParser class works like the Imageloader class: it reads from categories.txt, which contains every category on openTDB with 50 questions or more and the category's IDs (e.g., All is 0), and puts them into an object that acts as an associative array/hash/dictionary. API Requester now has a 'categories' data field that defaults to 0 (All), and no longer uses just easy difficulty. Use setCategory in the API Requester to change the category.